+++ NEWS.md
@@ -1,5 +1,8 @@
 # conflicted 1.0.0.9000
 
+* `.conflicts` environment is correctly removed and replaced each time
+  a new package is loaded (#28).
+
 # conflicted 1.0.0
 
 ### New functions